Factored verification code out of DLCClient and into DLCSignatureVerifier
